Added scriptable object.  Upgraded to .NET 4.  Added Velocity.  Increased memory to 2GB

Related work items: #2361

+++ Assets/_Scripts/TextUtils.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public static class TextUtils
 {
-    public static string GetPrettyNumber(float totalCount)
+    public static string GetPrettyNumber(double totalCount)
     {
         if (totalCount > 100000)
         {
